app/code/requirements.txt
awslabs-mcp-lambda-handler
pytz